// src/model/TrackData.ts
export  interface TrackData {
    deviceId: string;
    time: Date;
    lat: number;
    lng: number ;
    no2: number ;
    co: number ;
    o3: number ;
    so2: number ;
    pm1: number ;
    pm2_5: number ;
    pm10: number ;
    temp: number ;
    humi: number ;
    flog: number ;
    voc: number ;
}